# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "ChannelRequest": {
        "properties": {
          "url": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "ChannelRequest",
        "type": "object"
      },
      "ChannelResponse": {
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_offset": {
            "default": 0,
            "title": "Next Offset",
            "type": "integer"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/VideoItem"
            },
            "title": "Videos",
            "type": "array"
          }
        },
        "required": [
          "videos",
          "next_offset",
          "has_more"
        ],
        "title": "ChannelResponse",
        "type": "object"
      },
      "Clip": {
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "end": {
            "title": "End",
            "type": "string"
          },
          "order": {
            "default": 0,
            "title": "Order",
            "type": "integer"
          },
          "start": {
            "title": "Start",
            "type": "string"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "title": {
            "default": "Untitled",
            "title": "Title",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "view_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          }
        },
        "required": [
          "video_id",
          "start",
          "end"
        ],
        "title": "Clip",
        "type": "object"
      },
      "CompilationEntry": {
        "description": "One finished compilation in the usage log.",
        "properties": {
          "clips": {
            "title": "Clips",
            "type": "integer"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "integer"
          },
          "made": {
            "title": "Made",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "made",
          "clips",
          "duration_seconds"
        ],
        "title": "CompilationEntry",
        "type": "object"
      },
      "CompilationRequest": {
        "properties": {
          "clips": {
            "items": {
              "$ref": "#/components/schemas/Clip"
            },
            "minItems": 1,
            "title": "Clips",
            "type": "array"
          },
          "output_4k": {
            "default": false,
            "title": "Output 4K",
            "type": "boolean"
          }
        },
        "required": [
          "clips"
        ],
        "title": "CompilationRequest",
        "type": "object"
      },
      "DownloadRequest": {
        "properties": {
          "directory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directory"
          }
        },
        "title": "DownloadRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatus": {
        "properties": {
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "file_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Url"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "title": "Progress",
            "type": "string"
          },
          "saved_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Path"
          },
          "status": {
            "enum": [
              "queued",
              "downloading",
              "concatenating",
              "ready",
              "saved",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "progress",
          "error",
          "download_url",
          "file_url",
          "saved_path"
        ],
        "title": "JobStatus",
        "type": "object"
      },
      "UsageLogs": {
        "description": "The two tables shown on the Logs screen.",
        "properties": {
          "compilations": {
            "items": {
              "$ref": "#/components/schemas/CompilationEntry"
            },
            "title": "Compilations",
            "type": "array"
          },
          "videos": {
            "items": {
              "$ref": "#/components/schemas/UsageVideo"
            },
            "title": "Videos",
            "type": "array"
          }
        },
        "required": [
          "videos",
          "compilations"
        ],
        "title": "UsageLogs",
        "type": "object"
      },
      "UsageVideo": {
        "description": "How often one video has been included, plus the last title seen.",
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "last_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "view_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          }
        },
        "required": [
          "video_id",
          "title",
          "channel",
          "view_count",
          "duration_seconds",
          "thumbnail",
          "last_used",
          "count"
        ],
        "title": "UsageVideo",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoItem": {
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "compilation_count": {
            "default": 0,
            "title": "Compilation Count",
            "type": "integer"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "video_id": {
            "title": "Video Id",
            "type": "string"
          },
          "view_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Count"
          }
        },
        "required": [
          "video_id",
          "title",
          "thumbnail",
          "duration_seconds",
          "url",
          "compilation_count",
          "channel",
          "view_count"
        ],
        "title": "VideoItem",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CompCreator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/channels": {
      "post": {
        "description": "Return one page of videos for the channel in ``body``.",
        "operationId": "read_channel_api_channels_post",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Channel"
      }
    },
    "/api/compilations": {
      "post": {
        "description": "Start an export and return the queued job.",
        "operationId": "create_compilation_api_compilations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompilationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Compilation"
      }
    },
    "/api/compilations/{job_id}": {
      "get": {
        "description": "Return the current status of ``job_id``.",
        "operationId": "compilation_status_api_compilations__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compilation Status"
      }
    },
    "/api/compilations/{job_id}/cancel": {
      "post": {
        "description": "Stop ``job_id`` when it is still queued, downloading, or joining.",
        "operationId": "cancel_compilation_api_compilations__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Compilation"
      }
    },
    "/api/compilations/{job_id}/download": {
      "post": {
        "description": "Copy the finished video to Downloads or a chosen folder, then delete the job.",
        "operationId": "download_compilation_api_compilations__job_id__download_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DownloadRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Compilation"
      }
    },
    "/api/compilations/{job_id}/file": {
      "get": {
        "description": "Send the finished video to the browser, then delete the job after a grace period.\n\nA full response marks the job saved and keeps the file for\n``BROWSER_DOWNLOAD_GRACE_SECONDS`` so an interrupted download can be retried.",
        "operationId": "compilation_file_api_compilations__job_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compilation File"
      }
    },
    "/api/download-folder": {
      "get": {
        "description": "The default folder a finished video is saved into.",
        "operationId": "download_folder_api_download_folder_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Download Folder Api Download Folder Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Download Folder"
      }
    },
    "/api/logs": {
      "get": {
        "description": "Videos and compilations recorded on this computer.",
        "operationId": "read_logs_api_logs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageLogs"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Logs"
      }
    },
    "/api/session": {
      "get": {
        "description": "Answer 200 when the request may use the API, so the UI can ask for a password first.",
        "operationId": "session_api_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Session Api Session Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Session"
      }
    },
    "/api/usage": {
      "get": {
        "description": "Compilation counts for the comma-separated video ids in ``ids``.",
        "operationId": "read_usage_api_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "additionalProperties": {
                      "type": "integer"
                    },
                    "type": "object"
                  },
                  "title": "Response Read Usage Api Usage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Usage"
      }
    },
    "/health": {
      "get": {
        "description": "Report that the API process is accepting requests.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}